Parameters are now stored as float, likelihoods as double.
Generally, the parameters should not need 64 bit precision.
The idea is to save some space.  However, since likelihoods may
have large additive constants (since normalization is hard),
those are kept as double.

+++ examples/rosenbrock_example.cc
@@ -178,7 +178,7 @@ int main(int argc, char** argv) {
     rd.doSteps(rd.getNSteps(), nburn);
     
     if (rank == 0) {
-      std::vector<double> accept;
+      std::vector<float> accept;
       rd.getAcceptanceFrac(accept);
       double mnacc = accept[0];
       for (unsigned int i = 1; i < nwalkers; ++i)
@@ -186,7 +186,7 @@ int main(int argc, char** argv) {
       std::cout << "Mean acceptance: " << mnacc / static_cast<double>(nwalkers)
 		<< std::endl;
 
-      std::vector<double> acor;
+      std::vector<float> acor;
       bool succ = rd.getAcor(acor);
       if (succ) std::cout << "Autocorrelation length: " << acor[0] 
 			  << " " << acor[1] << std::endl;
